Fake randomize starting worker in DefaultJobPublisher based on Environment.TickCount
This is just to avoid always publishing to a first worker
since the new instance of publisher is created on each job batch
for scheduled/periodic workers.

Actually, immediate workers will also create one instance of publisher per batch
to inject into jobs if jobs require it in their constructors.