Fix retrying of jobs on error
This bug was introduced when we decided not to run scheduled jobs
directly in the scheduled workers, but to re-publish it to be executed
by immediate workers. The problem is, this republish put the job back
in the error queue (because it has Retries value set).

Now, we have a new field boolean "HasErrors", which is not sent
to the RabbitMq server - it is only used to route the message to the error
queue on publish.