Implement possibility to abort the entire batch from a job
This will allow a job to abort the entire batch. This is useful
when an application is executing jobs in one db transaction per batch.

For example, a job could accept a scoped DbContext subclass that
starts a transaction in DbContext subclass' constructor. On Dispose,
the DbContext subclass would Commit the transaction (if there is one,
i.e. not rolled back).

The Perform method could catch exceptions and decide to rollback the
db transaction. After db rollback the Perform method should throw
"BatchRolledBackException" with the inner exception set to the
original exception
.